+++ leshan-core/src/main/java/org/eclipse/leshan/core/request/SendRequest.java
@@ -18,13 +18,14 @@ package org.eclipse.leshan.core.request;
 import java.util.Map;
 import java.util.Map.Entry;
 
-import org.eclipse.leshan.core.node.TimestampedLwM2mNodes;
 import org.eclipse.leshan.core.node.LwM2mNode;
 import org.eclipse.leshan.core.node.LwM2mObject;
 import org.eclipse.leshan.core.node.LwM2mObjectInstance;
 import org.eclipse.leshan.core.node.LwM2mPath;
 import org.eclipse.leshan.core.node.LwM2mResourceInstance;
 import org.eclipse.leshan.core.node.LwM2mSingleResource;
+import org.eclipse.leshan.core.node.TimestampedLwM2mNodes;
+import org.eclipse.leshan.core.node.TimestampedLwM2mNodesImpl;
 import org.eclipse.leshan.core.request.exception.InvalidRequestException;
 import org.eclipse.leshan.core.response.SendResponse;
 import org.eclipse.leshan.core.util.Validate;
@@ -66,7 +67,7 @@ public class SendRequest implements UplinkRequest<SendResponse> {
     }
 
     public SendRequest(ContentFormat format, Map<LwM2mPath, LwM2mNode> nodes, Object coapRequest) {
-        data = new TimestampedLwM2mNodes(nodes);
+        data = new TimestampedLwM2mNodesImpl(nodes);
         // Validate Format
         if (format == null || !(format.equals(ContentFormat.SENML_CBOR) || format.equals(ContentFormat.SENML_JSON))) {
             throw new InvalidRequestException("Content format MUST be SenML_CBOR or SenML_JSON but was " + format);
